refactor: distribute testcases to seperate <testcase>_test.go files
