--- test/ngsetup.go
@@ -1,58 +1,0 @@
-package test
-
-import (
-	"fmt"
-	"net"
-
-	"git.cs.nctu.edu.tw/calee/sctp"
-	"github.com/calee0219/fatal"
-)
-
-const ngapPPID uint32 = 0x3c000000
-
-func getNgapIp(amfIP, ranIP string, amfPort, ranPort int) (amfAddr, ranAddr *sctp.SCTPAddr, err error) {
-	ips := []net.IPAddr{}
-	if ip, err1 := net.ResolveIPAddr("ip", amfIP); err1 != nil {
-		err = fmt.Errorf("Error resolving address '%s': %v", amfIP, err1)
-		return nil, nil, err
-	} else {
-		ips = append(ips, *ip)
-	}
-	amfAddr = &sctp.SCTPAddr{
-		IPAddrs: ips,
-		Port:    amfPort,
-	}
-	ips = []net.IPAddr{}
-	if ip, err1 := net.ResolveIPAddr("ip", ranIP); err1 != nil {
-		err = fmt.Errorf("Error resolving address '%s': %v", ranIP, err1)
-		return nil, nil, err
-	} else {
-		ips = append(ips, *ip)
-	}
-	ranAddr = &sctp.SCTPAddr{
-		IPAddrs: ips,
-		Port:    ranPort,
-	}
-	return amfAddr, ranAddr, nil
-}
-
-func ConnectToAmf(amfIP, ranIP string, amfPort, ranPort int) (*sctp.SCTPConn, error) {
-	amfAddr, ranAddr, err := getNgapIp(amfIP, ranIP, amfPort, ranPort)
-	if err != nil {
-		return nil, err
-	}
-	conn, err := sctp.DialSCTP("sctp", ranAddr, amfAddr)
-	if err != nil {
-		return nil, err
-	}
-	info, err := conn.GetDefaultSentParam()
-	if err != nil {
-		fatal.Fatalf("conn GetDefaultSentParam error in ConnectToAmf: %+v", err)
-	}
-	info.PPID = ngapPPID
-	err = conn.SetDefaultSentParam(info)
-	if err != nil {
-		return nil, err
-	}
-	return conn, nil
-}
