+++ test/pduSession_test.go
@@ -1,21 +1,12 @@
 package test
 
 import (
-	"github.com/free5gc/test/consumerTestdata/UDM/TestGenAuthData"
-	"github.com/free5gc/test/nasTestpacket"
-
 	"encoding/hex"
+	"fmt"
 	"net"
 	"testing"
 	"time"
 
-	"git.cs.nctu.edu.tw/calee/sctp"
-	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
-	"golang.org/x/net/icmp"
-	"golang.org/x/net/ipv4"
-
-	// ausf_context "github.com/free5gc/ausf/context"
 	"github.com/free5gc/nas"
 	"github.com/free5gc/nas/nasMessage"
 	"github.com/free5gc/nas/nasType"
@@ -23,242 +14,22 @@ import (
 	"github.com/free5gc/ngap"
 	"github.com/free5gc/ngap/ngapType"
 	"github.com/free5gc/openapi/models"
+	"github.com/free5gc/test/consumerTestdata/UDM/TestGenAuthData"
+	"github.com/free5gc/test/nasTestpacket"
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
+	"golang.org/x/net/icmp"
+	"golang.org/x/net/ipv4"
 )
 
-func recvUeConfigUpdateCmd(t *testing.T, recvMsg []byte, conn *sctp.SCTPConn) {
-	n, err := conn.Read(recvMsg)
-	assert.Nil(t, err)
-	ngapPdu, err := ngap.Decoder(recvMsg[:n])
-	assert.Nil(t, err)
-	assert.Equal(t, ngapPdu.Present, ngapType.NGAPPDUPresentInitiatingMessage, "Not NGAPPDUPresentInitiatingMessage")
-	assert.Equal(t, ngapPdu.InitiatingMessage.ProcedureCode.Value, ngapType.ProcedureCodeDownlinkNASTransport,
-		"Not ProcedureCodeDownlinkNASTransport")
-}
-
-// Registration
-func TestRegistration(t *testing.T) {
+// Registration -> Pdu Session Establishment -> Pdu Session Release
+func TestPDUSessionReleaseRequest(t *testing.T) {
 	var n int
 	var sendMsg []byte
 	var recvMsg = make([]byte, 2048)
 
 	// RAN connect to AMF
 	conn, err := ConnectToAmf(amfN2Ipv4Addr, ranN2Ipv4Addr, 38412, 9487)
-	assert.Nil(t, err)
-
-	// RAN connect to UPF
-	upfConn, err := ConnectToUpf(ranN3Ipv4Addr, upfN3Ipv4Addr, 2152, 2152)
-	assert.Nil(t, err)
-
-	// send NGSetupRequest Msg
-	sendMsg, err = GetNGSetupRequest([]byte("\x00\x01\x02"), 24, "free5gc")
-	assert.Nil(t, err)
-	_, err = conn.Write(sendMsg)
-	assert.Nil(t, err)
-
-	// receive NGSetupResponse Msg
-	n, err = conn.Read(recvMsg)
-	assert.Nil(t, err)
-	ngapPdu, err := ngap.Decoder(recvMsg[:n])
-	assert.Nil(t, err)
-	assert.True(t, ngapPdu.Present == ngapType.NGAPPDUPresentSuccessfulOutcome && ngapPdu.SuccessfulOutcome.ProcedureCode.Value == ngapType.ProcedureCodeNGSetup, "No NGSetupResponse received.")
-
-	// New UE
-	// ue := test.NewRanUeContext("imsi-2089300007487", 1, security.AlgCiphering128NEA2, security.AlgIntegrity128NIA2, models.AccessType__3_GPP_ACCESS)
-	ue := NewRanUeContext("imsi-2089300007487", 1, security.AlgCiphering128NEA0, security.AlgIntegrity128NIA2,
-		models.AccessType__3_GPP_ACCESS)
-	ue.AmfUeNgapId = 1
-	ue.AuthenticationSubs = GetAuthSubscription(TestGenAuthData.MilenageTestSet19.K,
-		TestGenAuthData.MilenageTestSet19.OPC,
-		TestGenAuthData.MilenageTestSet19.OP)
-
-	// insert UE data to MongoDB
-	servingPlmnId := "20893"
-	insertUeToMongoDB(t, ue, servingPlmnId)
-
-	// send InitialUeMessage(Registration Request)(imsi-2089300007487)
-	mobileIdentity5GS := nasType.MobileIdentity5GS{
-		Len:    12, // suci
-		Buffer: []uint8{0x01, 0x02, 0xf8, 0x39, 0xf0, 0xff, 0x00, 0x00, 0x00, 0x00, 0x47, 0x78},
-	}
-
-	ueSecurityCapability := ue.GetUESecurityCapability()
-	registrationRequest := nasTestpacket.GetRegistrationRequest(
-		nasMessage.RegistrationType5GSInitialRegistration, mobileIdentity5GS, nil, ueSecurityCapability, nil, nil, nil)
-	sendMsg, err = GetInitialUEMessage(ue.RanUeNgapId, registrationRequest, "")
-	assert.Nil(t, err)
-	_, err = conn.Write(sendMsg)
-	assert.Nil(t, err)
-
-	// receive NAS Authentication Request Msg
-	n, err = conn.Read(recvMsg)
-	assert.Nil(t, err)
-	ngapPdu, err = ngap.Decoder(recvMsg[:n])
-	assert.Nil(t, err)
-	assert.True(t, ngapPdu.Present == ngapType.NGAPPDUPresentInitiatingMessage, "No NGAP Initiating Message received.")
-
-	// Calculate for RES*
-	nasPdu := GetNasPdu(ue, ngapPdu.InitiatingMessage.Value.DownlinkNASTransport)
-	require.NotNil(t, nasPdu)
-	require.NotNil(t, nasPdu.GmmMessage, "GMM message is nil")
-	require.Equal(t, nasPdu.GmmHeader.GetMessageType(), nas.MsgTypeAuthenticationRequest,
-		"Received wrong GMM message. Expected Authentication Request.")
-	rand := nasPdu.AuthenticationRequest.GetRANDValue()
-	resStat := ue.DeriveRESstarAndSetKey(ue.AuthenticationSubs, rand[:], "5G:mnc093.mcc208.3gppnetwork.org")
-
-	// send NAS Authentication Response
-	pdu := nasTestpacket.GetAuthenticationResponse(resStat, "")
-	sendMsg, err = GetUplinkNASTransport(ue.AmfUeNgapId, ue.RanUeNgapId, pdu)
-	assert.Nil(t, err)
-	_, err = conn.Write(sendMsg)
-	assert.Nil(t, err)
-
-	// receive NAS Security Mode Command Msg
-	n, err = conn.Read(recvMsg)
-	assert.Nil(t, err)
-	ngapPdu, err = ngap.Decoder(recvMsg[:n])
-	assert.Nil(t, err)
-	assert.NotNil(t, ngapPdu)
-	nasPdu = GetNasPdu(ue, ngapPdu.InitiatingMessage.Value.DownlinkNASTransport)
-	require.NotNil(t, nasPdu)
-	require.NotNil(t, nasPdu.GmmMessage, "GMM message is nil")
-	require.Equal(t, nasPdu.GmmHeader.GetMessageType(), nas.MsgTypeSecurityModeCommand,
-		"Received wrong GMM message. Expected Security Mode Command.")
-
-	// send NAS Security Mode Complete Msg
-	registrationRequestWith5GMM := nasTestpacket.GetRegistrationRequest(nasMessage.RegistrationType5GSInitialRegistration,
-		mobileIdentity5GS, nil, ueSecurityCapability, ue.Get5GMMCapability(), nil, nil)
-	pdu = nasTestpacket.GetSecurityModeComplete(registrationRequestWith5GMM)
-	pdu, err = EncodeNasPduWithSecurity(ue, pdu, nas.SecurityHeaderTypeIntegrityProtectedAndCipheredWithNew5gNasSecurityContext, true, true)
-	assert.Nil(t, err)
-	sendMsg, err = GetUplinkNASTransport(ue.AmfUeNgapId, ue.RanUeNgapId, pdu)
-	assert.Nil(t, err)
-	_, err = conn.Write(sendMsg)
-	assert.Nil(t, err)
-
-	// receive ngap Initial Context Setup Request Msg
-	n, err = conn.Read(recvMsg)
-	assert.Nil(t, err)
-	ngapPdu, err = ngap.Decoder(recvMsg[:n])
-	assert.Nil(t, err)
-	assert.True(t, ngapPdu.Present == ngapType.NGAPPDUPresentInitiatingMessage &&
-		ngapPdu.InitiatingMessage.ProcedureCode.Value == ngapType.ProcedureCodeInitialContextSetup,
-		"No InitialContextSetup received.")
-
-	// send ngap Initial Context Setup Response Msg
-	sendMsg, err = GetInitialContextSetupResponse(ue.AmfUeNgapId, ue.RanUeNgapId)
-	assert.Nil(t, err)
-	_, err = conn.Write(sendMsg)
-	assert.Nil(t, err)
-
-	// send NAS Registration Complete Msg
-	pdu = nasTestpacket.GetRegistrationComplete(nil)
-	pdu, err = EncodeNasPduWithSecurity(ue, pdu, nas.SecurityHeaderTypeIntegrityProtectedAndCiphered, true, false)
-	assert.Nil(t, err)
-	sendMsg, err = GetUplinkNASTransport(ue.AmfUeNgapId, ue.RanUeNgapId, pdu)
-	assert.Nil(t, err)
-	_, err = conn.Write(sendMsg)
-	assert.Nil(t, err)
-
-	// receive UE Configuration Update Command Msg
-	recvUeConfigUpdateCmd(t, recvMsg, conn)
-
-	time.Sleep(100 * time.Millisecond)
-
-	// send GetPduSessionEstablishmentRequest Msg
-	sNssai := models.Snssai{
-		Sst: 1,
-		Sd:  "fedcba",
-	}
-	pdu = nasTestpacket.GetUlNasTransport_PduSessionEstablishmentRequest(10, nasMessage.ULNASTransportRequestTypeInitialRequest, "internet", &sNssai)
-	pdu, err = EncodeNasPduWithSecurity(ue, pdu, nas.SecurityHeaderTypeIntegrityProtectedAndCiphered, true, false)
-	assert.Nil(t, err)
-	sendMsg, err = GetUplinkNASTransport(ue.AmfUeNgapId, ue.RanUeNgapId, pdu)
-	assert.Nil(t, err)
-	_, err = conn.Write(sendMsg)
-	assert.Nil(t, err)
-
-	// receive 12. NGAP-PDU Session Resource Setup Request(DL nas transport((NAS msg-PDU session setup Accept)))
-	n, err = conn.Read(recvMsg)
-	assert.Nil(t, err)
-	ngapPdu, err = ngap.Decoder(recvMsg[:n])
-	assert.Nil(t, err)
-	assert.True(t, ngapPdu.Present == ngapType.NGAPPDUPresentInitiatingMessage &&
-		ngapPdu.InitiatingMessage.ProcedureCode.Value == ngapType.ProcedureCodePDUSessionResourceSetup,
-		"No PDUSessionResourceSetup received.")
-
-	// send 14. NGAP-PDU Session Resource Setup Response
-	sendMsg, err = GetPDUSessionResourceSetupResponse(10, ue.AmfUeNgapId, ue.RanUeNgapId, ranN3Ipv4Addr)
-	assert.Nil(t, err)
-	_, err = conn.Write(sendMsg)
-	assert.Nil(t, err)
-
-	// wait 1s
-	time.Sleep(1 * time.Second)
-
-	// Send the dummy packet
-	// ping IP(tunnel IP) from 10.60.0.2(127.0.0.1) to 10.60.0.20(127.0.0.8)
-	gtpHdr, err := hex.DecodeString("32ff00340000000100000000")
-	assert.Nil(t, err)
-	icmpData, err := hex.DecodeString("8c870d0000000000101112131415161718191a1b1c1d1e1f202122232425262728292a2b2c2d2e2f3031323334353637")
-	assert.Nil(t, err)
-
-	ipv4hdr := ipv4.Header{
-		Version:  4,
-		Len:      20,
-		Protocol: 1,
-		Flags:    0,
-		TotalLen: 48,
-		TTL:      64,
-		Src:      net.ParseIP("10.60.0.1").To4(),
-		Dst:      net.ParseIP("10.60.0.101").To4(),
-		ID:       1,
-	}
-	checksum := CalculateIpv4HeaderChecksum(&ipv4hdr)
-	ipv4hdr.Checksum = int(checksum)
-
-	v4HdrBuf, err := ipv4hdr.Marshal()
-	assert.Nil(t, err)
-	tt := append(gtpHdr, v4HdrBuf...)
-
-	m := icmp.Message{
-		Type: ipv4.ICMPTypeEcho, Code: 0,
-		Body: &icmp.Echo{
-			ID: 12394, Seq: 1,
-			Data: icmpData,
-		},
-	}
-	b, err := m.Marshal(nil)
-	assert.Nil(t, err)
-	b[2] = 0xaf
-	b[3] = 0x88
-	_, err = upfConn.Write(append(tt, b...))
-	assert.Nil(t, err)
-
-	time.Sleep(1 * time.Second)
-
-	// delete test data
-	DelAuthSubscriptionToMongoDB(ue.Supi)
-	DelAccessAndMobilitySubscriptionDataFromMongoDB(ue.Supi, servingPlmnId)
-	DelSmfSelectionSubscriptionDataFromMongoDB(ue.Supi, servingPlmnId)
-
-	// close Connection
-	conn.Close()
-
-	// terminate all NF
-	NfTerminate()
-}
-
-func TestDuplicateRegistration(t *testing.T) {
-	var n int
-	var sendMsg []byte
-	var recvMsg = make([]byte, 2048)
-
-	// RAN connect to AMF
-	conn, err := ConnectToAmf(amfN2Ipv4Addr, ranN2Ipv4Addr, 38412, 9487)
-	assert.Nil(t, err)
-
-	// RAN connect to UPF
-	upfConn, err := ConnectToUpf(ranN3Ipv4Addr, "10.200.200.102", 2152, 2152)
 	assert.Nil(t, err)
 
 	// send NGSetupRequest Msg
@@ -276,7 +47,6 @@ func TestDuplicateRegistration(t *testing.T) {
 	// New UE
 	ue := NewRanUeContext("imsi-2089300007487", 1, security.AlgCiphering128NEA0, security.AlgIntegrity128NIA2,
 		models.AccessType__3_GPP_ACCESS)
-	// ue := NewRanUeContext("imsi-2089300007487", 1, security.AlgCiphering128NEA0, security.AlgIntegrity128NIA0, models.AccessType__3_GPP_ACCESS)
 	ue.AmfUeNgapId = 1
 	ue.AuthenticationSubs = GetAuthSubscription(TestGenAuthData.MilenageTestSet19.K,
 		TestGenAuthData.MilenageTestSet19.OPC,
@@ -289,7 +59,7 @@ func TestDuplicateRegistration(t *testing.T) {
 	// send InitialUeMessage(Registration Request)(imsi-2089300007487)
 	mobileIdentity5GS := nasType.MobileIdentity5GS{
 		Len:    12, // suci
-		Buffer: []uint8{0x01, 0x02, 0xf8, 0x39, 0xf0, 0xff, 0x00, 0x00, 0x00, 0x00, 0x47, 0x78},
+		Buffer: []uint8{0x01, 0x02, 0xf8, 0x39, 0xf0, 0x00, 0x00, 0x00, 0x00, 0x00, 0x47, 0x78},
 	}
 	ueSecurityCapability := ue.GetUESecurityCapability()
 	registrationRequest := nasTestpacket.GetRegistrationRequest(
@@ -368,9 +138,7 @@ func TestDuplicateRegistration(t *testing.T) {
 	// receive UE Configuration Update Command Msg
 	recvUeConfigUpdateCmd(t, recvMsg, conn)
 
-	time.Sleep(100 * time.Millisecond)
-
-	// send GetPduSessionEstablishmentRequest Msg
+	// send PduSessionEstablishmentRequest Msg
 	sNssai := models.Snssai{
 		Sst: 1,
 		Sd:  "fedcba",
@@ -378,9 +146,9 @@ func TestDuplicateRegistration(t *testing.T) {
 	pdu = nasTestpacket.GetUlNasTransport_PduSessionEstablishmentRequest(10, nasMessage.ULNASTransportRequestTypeInitialRequest, "internet", &sNssai)
 	pdu, err = EncodeNasPduWithSecurity(ue, pdu, nas.SecurityHeaderTypeIntegrityProtectedAndCiphered, true, false)
 	assert.Nil(t, err)
-	PduSessionEstablishmentRequestMSG, err := GetUplinkNASTransport(ue.AmfUeNgapId, ue.RanUeNgapId, pdu)
+	sendMsg, err = GetUplinkNASTransport(ue.AmfUeNgapId, ue.RanUeNgapId, pdu)
 	assert.Nil(t, err)
-	_, err = conn.Write(PduSessionEstablishmentRequestMSG)
+	_, err = conn.Write(sendMsg)
 	assert.Nil(t, err)
 
 	// receive 12. NGAP-PDU Session Resource Setup Request(DL nas transport((NAS msg-PDU session setup Accept)))
@@ -395,81 +163,42 @@ func TestDuplicateRegistration(t *testing.T) {
 	_, err = conn.Write(sendMsg)
 	assert.Nil(t, err)
 
-	// wait 1s
-	time.Sleep(1 * time.Second)
-
-	//send PduSessionEstablishment Request Msg again to make error
-
-	_, err = conn.Write(PduSessionEstablishmentRequestMSG)
+	// Send Pdu Session Establishment Release Request
+	pdu = nasTestpacket.GetUlNasTransport_PduSessionReleaseRequest(10)
+	pdu, err = EncodeNasPduWithSecurity(ue, pdu, nas.SecurityHeaderTypeIntegrityProtectedAndCiphered, true, false)
+	assert.Nil(t, err)
+	sendMsg, err = GetUplinkNASTransport(ue.AmfUeNgapId, ue.RanUeNgapId, pdu)
+	assert.Nil(t, err)
+	_, err = conn.Write(sendMsg)
 	assert.Nil(t, err)
 
+	time.Sleep(1000 * time.Millisecond)
+	// send N2 Resource Release Ack(PDUSession Resource Release Response)
 	sendMsg, err = GetPDUSessionResourceReleaseResponse(ue.AmfUeNgapId, ue.RanUeNgapId)
 	assert.Nil(t, err)
 	_, err = conn.Write(sendMsg)
 	assert.Nil(t, err)
 
-	// receive 12. NGAP-PDU Session Resource Setup Request(DL nas transport((NAS msg-PDU session setup Accept)))
-	n, err = conn.Read(recvMsg)
-	assert.Nil(t, err)
-	_, err = ngap.Decoder(recvMsg[:n])
-	assert.Nil(t, err)
+	// wait 10 ms
+	time.Sleep(1000 * time.Millisecond)
 
-	time.Sleep(1 * time.Second)
-
-	// send 14. NGAP-PDU Session Resource Setup Response
-	sendMsg, err = GetPDUSessionResourceSetupResponse(10, ue.AmfUeNgapId, ue.RanUeNgapId, ranN3Ipv4Addr)
+	//send N1 PDU Session Release Ack PDU session release complete
+	pdu = nasTestpacket.GetUlNasTransport_PduSessionReleaseComplete(10, nasMessage.ULNASTransportRequestTypeExistingPduSession, "internet", &sNssai)
+	pdu, err = EncodeNasPduWithSecurity(ue, pdu, nas.SecurityHeaderTypeIntegrityProtectedAndCiphered, true, false)
+	assert.Nil(t, err)
+	sendMsg, err = GetUplinkNASTransport(ue.AmfUeNgapId, ue.RanUeNgapId, pdu)
 	assert.Nil(t, err)
 	_, err = conn.Write(sendMsg)
 	assert.Nil(t, err)
 
-	// wait 1s
+	// wait result
 	time.Sleep(1 * time.Second)
-
-	// Send the dummy packet
-	// ping IP(tunnel IP) from 10.60.0.2(127.0.0.1) to 10.60.0.20(127.0.0.8)
-	gtpHdr, err := hex.DecodeString("32ff00340000000300000000")
-	assert.Nil(t, err)
-	icmpData, err := hex.DecodeString("8c870d0000000000101112131415161718191a1b1c1d1e1f202122232425262728292a2b2c2d2e2f3031323334353637")
-	assert.Nil(t, err)
-
-	ipv4hdr := ipv4.Header{
-		Version:  4,
-		Len:      20,
-		Protocol: 1,
-		Flags:    0,
-		TotalLen: 48,
-		TTL:      64,
-		Src:      net.ParseIP("10.60.0.2").To4(),
-		Dst:      net.ParseIP("10.60.0.101").To4(),
-		ID:       1,
-	}
-	checksum := CalculateIpv4HeaderChecksum(&ipv4hdr)
-	ipv4hdr.Checksum = int(checksum)
-
-	v4HdrBuf, err := ipv4hdr.Marshal()
-	assert.Nil(t, err)
-	tt := append(gtpHdr, v4HdrBuf...)
-	assert.Nil(t, err)
-
-	m := icmp.Message{
-		Type: ipv4.ICMPTypeEcho, Code: 0,
-		Body: &icmp.Echo{
-			ID: 12394, Seq: 1,
-			Data: icmpData,
-		},
-	}
-	b, err := m.Marshal(nil)
-	assert.Nil(t, err)
-	b[2] = 0xaf
-	b[3] = 0x88
-	_, err = upfConn.Write(append(tt, b...))
-	assert.Nil(t, err)
 
 	// delete test data
 	DelAuthSubscriptionToMongoDB(ue.Supi)
 	DelAccessAndMobilitySubscriptionDataFromMongoDB(ue.Supi, servingPlmnId)
 	DelSmfSelectionSubscriptionDataFromMongoDB(ue.Supi, servingPlmnId)
-	time.Sleep(1 * time.Second)
+
 	// close Connection
 	conn.Close()
 
@@ -477,15 +206,252 @@ func TestDuplicateRegistration(t *testing.T) {
 	NfTerminate()
 }
 
-func TestMultiAmfRegistration(t *testing.T) {
-	time.Sleep(3 * time.Second)
-
+func TestPDUSessionReleaseAbnormal(t *testing.T) {
 	var n int
 	var sendMsg []byte
 	var recvMsg = make([]byte, 2048)
 
-	// RAN connect to old amf
+	// RAN connect to AMF
 	conn, err := ConnectToAmf(amfN2Ipv4Addr, ranN2Ipv4Addr, 38412, 9487)
+	assert.Nil(t, err)
+
+	// send NGSetupRequest Msg
+	sendMsg, err = GetNGSetupRequest([]byte("\x00\x01\x02"), 24, "free5gc")
+	assert.Nil(t, err)
+	_, err = conn.Write(sendMsg)
+	assert.Nil(t, err)
+
+	// receive NGSetupResponse Msg
+	n, err = conn.Read(recvMsg)
+	assert.Nil(t, err)
+	_, err = ngap.Decoder(recvMsg[:n])
+	assert.Nil(t, err)
+
+	// New UE
+	ue := NewRanUeContext("imsi-2089300007487", 1, security.AlgCiphering128NEA0, security.AlgIntegrity128NIA2,
+		models.AccessType__3_GPP_ACCESS)
+	ue.AmfUeNgapId = 1
+	ue.AuthenticationSubs = GetAuthSubscription(TestGenAuthData.MilenageTestSet19.K,
+		TestGenAuthData.MilenageTestSet19.OPC,
+		TestGenAuthData.MilenageTestSet19.OP)
+	// insert UE data to MongoDB
+
+	servingPlmnId := "20893"
+	InsertAuthSubscriptionToMongoDB(ue.Supi, ue.AuthenticationSubs)
+	getData := GetAuthSubscriptionFromMongoDB(ue.Supi)
+	assert.NotNil(t, getData)
+	{
+		amData := GetAccessAndMobilitySubscriptionData()
+		InsertAccessAndMobilitySubscriptionDataToMongoDB(ue.Supi, amData, servingPlmnId)
+		getData := GetAccessAndMobilitySubscriptionDataFromMongoDB(ue.Supi, servingPlmnId)
+		assert.NotNil(t, getData)
+	}
+	{
+		smfSelData := GetSmfSelectionSubscriptionData()
+		InsertSmfSelectionSubscriptionDataToMongoDB(ue.Supi, smfSelData, servingPlmnId)
+		getData := GetSmfSelectionSubscriptionDataFromMongoDB(ue.Supi, servingPlmnId)
+		assert.NotNil(t, getData)
+	}
+	{
+		smSelData := GetSessionManagementSubscriptionData()
+		InsertSessionManagementSubscriptionDataToMongoDB(ue.Supi, servingPlmnId, smSelData)
+		getData := GetSessionManagementDataFromMongoDB(ue.Supi, servingPlmnId)
+		assert.NotNil(t, getData)
+	}
+	{
+		amPolicyData := GetAmPolicyData()
+		InsertAmPolicyDataToMongoDB(ue.Supi, amPolicyData)
+		getData := GetAmPolicyDataFromMongoDB(ue.Supi)
+		assert.NotNil(t, getData)
+	}
+	{
+		smPolicyData := GetSmPolicyData()
+		InsertSmPolicyDataToMongoDB(ue.Supi, smPolicyData)
+		getData := GetSmPolicyDataFromMongoDB(ue.Supi)
+		assert.NotNil(t, getData)
+	}
+
+	// send InitialUeMessage(Registration Request)(imsi-2089300007487)
+	mobileIdentity5GS := nasType.MobileIdentity5GS{
+		Len:    12, // suci
+		Buffer: []uint8{0x01, 0x02, 0xf8, 0x39, 0xf0, 0x00, 0x00, 0x00, 0x00, 0x00, 0x47, 0x78},
+	}
+	ueSecurityCapability := ue.GetUESecurityCapability()
+	registrationRequest := nasTestpacket.GetRegistrationRequest(
+		nasMessage.RegistrationType5GSInitialRegistration, mobileIdentity5GS, nil, ueSecurityCapability, nil, nil, nil)
+	sendMsg, err = GetInitialUEMessage(ue.RanUeNgapId, registrationRequest, "")
+	assert.Nil(t, err)
+	_, err = conn.Write(sendMsg)
+	assert.Nil(t, err)
+
+	// receive NAS Authentication Request Msg
+	n, err = conn.Read(recvMsg)
+	assert.Nil(t, err)
+	ngapMsg, err := ngap.Decoder(recvMsg[:n])
+	assert.Nil(t, err)
+
+	// Calculate for RES*
+	nasPdu := GetNasPdu(ue, ngapMsg.InitiatingMessage.Value.DownlinkNASTransport)
+	require.NotNil(t, nasPdu)
+	require.NotNil(t, nasPdu.GmmMessage, "GMM message is nil")
+	require.Equal(t, nasPdu.GmmHeader.GetMessageType(), nas.MsgTypeAuthenticationRequest,
+		"Received wrong GMM message. Expected Authentication Request.")
+	rand := nasPdu.AuthenticationRequest.GetRANDValue()
+	resStat := ue.DeriveRESstarAndSetKey(ue.AuthenticationSubs, rand[:], "5G:mnc093.mcc208.3gppnetwork.org")
+
+	// send NAS Authentication Response
+	pdu := nasTestpacket.GetAuthenticationResponse(resStat, "")
+	sendMsg, err = GetUplinkNASTransport(ue.AmfUeNgapId, ue.RanUeNgapId, pdu)
+	assert.Nil(t, err)
+	_, err = conn.Write(sendMsg)
+	assert.Nil(t, err)
+
+	// receive NAS Security Mode Command Msg
+	n, err = conn.Read(recvMsg)
+	assert.Nil(t, err)
+	ngapPdu, err := ngap.Decoder(recvMsg[:n])
+	require.Nil(t, err)
+	require.NotNil(t, ngapPdu)
+	nasPdu = GetNasPdu(ue, ngapPdu.InitiatingMessage.Value.DownlinkNASTransport)
+	require.NotNil(t, nasPdu)
+	require.NotNil(t, nasPdu.GmmMessage, "GMM message is nil")
+	require.Equal(t, nasPdu.GmmHeader.GetMessageType(), nas.MsgTypeSecurityModeCommand,
+		"Received wrong GMM message. Expected Security Mode Command.")
+
+	// send NAS Security Mode Complete Msg
+	registrationRequestWith5GMM := nasTestpacket.GetRegistrationRequest(nasMessage.RegistrationType5GSInitialRegistration,
+		mobileIdentity5GS, nil, ueSecurityCapability, ue.Get5GMMCapability(), nil, nil)
+	pdu = nasTestpacket.GetSecurityModeComplete(registrationRequestWith5GMM)
+	pdu, err = EncodeNasPduWithSecurity(ue, pdu, nas.SecurityHeaderTypeIntegrityProtectedAndCipheredWithNew5gNasSecurityContext, true, true)
+	assert.Nil(t, err)
+	sendMsg, err = GetUplinkNASTransport(ue.AmfUeNgapId, ue.RanUeNgapId, pdu)
+	assert.Nil(t, err)
+	_, err = conn.Write(sendMsg)
+	assert.Nil(t, err)
+
+	// receive ngap Initial Context Setup Request Msg
+	n, err = conn.Read(recvMsg)
+	assert.Nil(t, err)
+	_, err = ngap.Decoder(recvMsg[:n])
+	assert.Nil(t, err)
+
+	// send ngap Initial Context Setup Response Msg
+	sendMsg, err = GetInitialContextSetupResponse(ue.AmfUeNgapId, ue.RanUeNgapId)
+	assert.Nil(t, err)
+	_, err = conn.Write(sendMsg)
+	assert.Nil(t, err)
+
+	// send NAS Registration Complete Msg
+	pdu = nasTestpacket.GetRegistrationComplete(nil)
+	pdu, err = EncodeNasPduWithSecurity(ue, pdu, nas.SecurityHeaderTypeIntegrityProtectedAndCiphered, true, false)
+	assert.Nil(t, err)
+	sendMsg, err = GetUplinkNASTransport(ue.AmfUeNgapId, ue.RanUeNgapId, pdu)
+	assert.Nil(t, err)
+	_, err = conn.Write(sendMsg)
+	assert.Nil(t, err)
+
+	// receive UE Configuration Update Command Msg
+	recvUeConfigUpdateCmd(t, recvMsg, conn)
+
+	// send PduSessionEstablishmentRequest Msg
+	sNssai := models.Snssai{
+		Sst: 1,
+		Sd:  "fedcba",
+	}
+	pdu = nasTestpacket.GetUlNasTransport_PduSessionEstablishmentRequest(10, nasMessage.ULNASTransportRequestTypeInitialRequest, "internet", &sNssai)
+	pdu, err = EncodeNasPduWithSecurity(ue, pdu, nas.SecurityHeaderTypeIntegrityProtectedAndCiphered, true, false)
+	assert.Nil(t, err)
+	sendMsg, err = GetUplinkNASTransport(ue.AmfUeNgapId, ue.RanUeNgapId, pdu)
+	assert.Nil(t, err)
+	_, err = conn.Write(sendMsg)
+	assert.Nil(t, err)
+
+	// receive 12. NGAP-PDU Session Resource Setup Request(DL nas transport((NAS msg-PDU session setup Accept)))
+	n, err = conn.Read(recvMsg)
+	assert.Nil(t, err)
+	_, err = ngap.Decoder(recvMsg[:n])
+	assert.Nil(t, err)
+
+	// send 14. NGAP-PDU Session Resource Setup Response
+	sendMsg, err = GetPDUSessionResourceSetupResponse(10, ue.AmfUeNgapId, ue.RanUeNgapId, ranN3Ipv4Addr)
+	assert.Nil(t, err)
+	_, err = conn.Write(sendMsg)
+	assert.Nil(t, err)
+
+	// Send Pdu Session Release Request
+	pdu = nasTestpacket.GetUlNasTransport_PduSessionReleaseRequest(10)
+	pdu, err = EncodeNasPduWithSecurity(ue, pdu, nas.SecurityHeaderTypeIntegrityProtectedAndCiphered, true, false)
+	assert.Nil(t, err)
+	sendMsg, err = GetUplinkNASTransport(ue.AmfUeNgapId, ue.RanUeNgapId, pdu)
+	assert.Nil(t, err)
+	_, err = conn.Write(sendMsg)
+	assert.Nil(t, err)
+
+	// receive PDU Session Resource Release Command
+	n, err = conn.Read(recvMsg)
+	assert.Nil(t, err)
+	_, err = ngap.Decoder(recvMsg[:n])
+	assert.Nil(t, err)
+
+	time.Sleep(1000 * time.Millisecond)
+	// send N2 Resource Release Ack(PDUSession Resource Release Response)
+	sendMsg, err = GetPDUSessionResourceReleaseResponse(ue.AmfUeNgapId, ue.RanUeNgapId)
+	assert.Nil(t, err)
+	_, err = conn.Write(sendMsg)
+	assert.Nil(t, err)
+
+	// send ngap UE Context Release Request
+	pduSessionIDList := []int64{10}
+	sendMsg, err = GetUEContextReleaseRequest(ue.AmfUeNgapId, ue.RanUeNgapId, pduSessionIDList)
+	assert.Nil(t, err)
+	_, err = conn.Write(sendMsg)
+	assert.Nil(t, err)
+
+	// receive UE Context Release Command
+	n, err = conn.Read(recvMsg)
+	assert.Nil(t, err)
+	_, err = ngap.Decoder(recvMsg[:n])
+	assert.Nil(t, err)
+
+	// send ngap UE Context Release Complete
+	sendMsg, err = GetUEContextReleaseComplete(ue.AmfUeNgapId, ue.RanUeNgapId, nil)
+	assert.Nil(t, err)
+	_, err = conn.Write(sendMsg)
+	assert.Nil(t, err)
+
+	// Just test registration no occur panic
+	// send InitialUeMessage(Registration Request)(imsi-2089300007487)
+	sendMsg, err = GetInitialUEMessage(ue.RanUeNgapId, registrationRequest, "")
+	assert.Nil(t, err)
+	_, err = conn.Write(sendMsg)
+	assert.Nil(t, err)
+
+	// wait result
+	time.Sleep(2 * time.Second)
+
+	// delete test data
+	DelAuthSubscriptionToMongoDB(ue.Supi)
+	DelAccessAndMobilitySubscriptionDataFromMongoDB(ue.Supi, servingPlmnId)
+	DelSmfSelectionSubscriptionDataFromMongoDB(ue.Supi, servingPlmnId)
+
+	// close Connection
+	conn.Close()
+
+	// terminate all NF
+	NfTerminate()
+}
+
+func TestRequestTwoPDUSessions(t *testing.T) {
+	var n int
+	var sendMsg []byte
+	var recvMsg = make([]byte, 2048)
+
+	// RAN connect to AMF
+	conn, err := ConnectToAmf(amfN2Ipv4Addr, ranN2Ipv4Addr, 38412, 9487)
+	assert.Nil(t, err)
+
+	// RAN connect to UPF
+	upfConn, err := ConnectToUpf(ranN3Ipv4Addr, "10.200.200.101", 3001, 2152)
 	assert.Nil(t, err)
 
 	// send NGSetupRequest Msg
@@ -501,33 +467,16 @@ func TestMultiAmfRegistration(t *testing.T) {
 	assert.Nil(t, err)
 	assert.True(t, ngapPdu.Present == ngapType.NGAPPDUPresentSuccessfulOutcome && ngapPdu.SuccessfulOutcome.ProcedureCode.Value == ngapType.ProcedureCodeNGSetup, "No NGSetupResponse received.")
 
-	// RAN connect to new amf
-	conn2, err := ConnectToAmf(amfN2Ipv4Addr2, ranN2Ipv4Addr, 38413, 9488)
-	assert.Nil(t, err)
-
-	// send NGSetupRequest Msg
-	sendMsg, err = GetNGSetupRequest([]byte("\x00\x01\x02"), 24, "free5gc")
-	assert.Nil(t, err)
-	_, err = conn2.Write(sendMsg)
-	assert.Nil(t, err)
-
-	// receive NGSetupResponse Msg
-	n, err = conn2.Read(recvMsg)
-	assert.Nil(t, err)
-	ngapPdu, err = ngap.Decoder(recvMsg[:n])
-	assert.Nil(t, err)
-	assert.True(t, ngapPdu.Present == ngapType.NGAPPDUPresentSuccessfulOutcome && ngapPdu.SuccessfulOutcome.ProcedureCode.Value == ngapType.ProcedureCodeNGSetup, "No NGSetupResponse received.")
-
 	// New UE
 	// ue := NewRanUeContext("imsi-2089300007487", 1, security.AlgCiphering128NEA2, security.AlgIntegrity128NIA2, models.AccessType__3_GPP_ACCESS)
-	ue := NewRanUeContext("imsi-2089300007487", 1, security.AlgCiphering128NEA2, security.AlgIntegrity128NIA2,
+	ue := NewRanUeContext("imsi-2089300007487", 1, security.AlgCiphering128NEA0, security.AlgIntegrity128NIA2,
 		models.AccessType__3_GPP_ACCESS)
 	ue.AmfUeNgapId = 1
 	ue.AuthenticationSubs = GetAuthSubscription(TestGenAuthData.MilenageTestSet19.K,
 		TestGenAuthData.MilenageTestSet19.OPC,
 		TestGenAuthData.MilenageTestSet19.OP)
-
 	// insert UE data to MongoDB
+
 	servingPlmnId := "20893"
 	insertUeToMongoDB(t, ue, servingPlmnId)
 
@@ -600,16 +549,6 @@ func TestMultiAmfRegistration(t *testing.T) {
 		ngapPdu.InitiatingMessage.ProcedureCode.Value == ngapType.ProcedureCodeInitialContextSetup,
 		"No InitialContextSetup received.")
 
-	var guti *nasType.GUTI5G
-	for _, ie := range ngapPdu.InitiatingMessage.Value.InitialContextSetupRequest.ProtocolIEs.List {
-		if ie.Id.Value == ngapType.ProtocolIEIDNASPDU {
-			payload := []byte(ie.Value.NASPDU.Value)
-			m, err := NASDecode(ue, nas.GetSecurityHeaderType(payload), payload)
-			assert.Nil(t, err)
-			guti = m.RegistrationAccept.GUTI5G
-		}
-	}
-
 	// send ngap Initial Context Setup Response Msg
 	sendMsg, err = GetInitialContextSetupResponse(ue.AmfUeNgapId, ue.RanUeNgapId)
 	assert.Nil(t, err)
@@ -625,55 +564,173 @@ func TestMultiAmfRegistration(t *testing.T) {
 	_, err = conn.Write(sendMsg)
 	assert.Nil(t, err)
 
-	time.Sleep(1 * time.Second)
+	time.Sleep(100 * time.Millisecond)
 
-	conn.Close()
+	// receive UE Configuration Update Command Msg
+	recvUeConfigUpdateCmd(t, recvMsg, conn)
 
-	// -----Second registration----- ue re-registration with GUTI
-
-	GUTI5GS := nasType.MobileIdentity5GS{
-		Iei:    guti.Iei,
-		Len:    guti.Len,
-		Buffer: guti.Octet[:],
+	// send GetPduSessionEstablishmentRequest Msg
+	sNssai := models.Snssai{
+		Sst: 1,
+		Sd:  "fedcba",
 	}
-
-	// Send registration request to new amf
-	innerRegistrationRequest := nasTestpacket.GetRegistrationRequest(nasMessage.RegistrationType5GSInitialRegistration,
-		GUTI5GS, nil, ue.GetUESecurityCapability(), ue.Get5GMMCapability(), nil, nil)
-	pdu, err = EncodeNasPduWithSecurity(ue, innerRegistrationRequest, nas.SecurityHeaderTypeIntegrityProtectedAndCiphered, true, false)
-	require.Nil(t, err)
-	registrationRequest = nasTestpacket.GetRegistrationRequest(nasMessage.RegistrationType5GSInitialRegistration,
-		GUTI5GS, nil, ueSecurityCapability, ue.Get5GMMCapability(), pdu, nil)
-	pdu, err = EncodeNasPduWithSecurity(ue, registrationRequest, nas.SecurityHeaderTypeIntegrityProtected, true, false)
-	require.Nil(t, err)
-	sendMsg, err = GetInitialUEMessage(ue.RanUeNgapId, pdu, "fe0000000001")
-	require.Nil(t, err)
-
-	_, err = conn2.Write(sendMsg)
-	require.Nil(t, err)
-
-	// receive ngap Initial Context Setup Request Msg
-	n, err = conn2.Read(recvMsg)
-	require.Nil(t, err)
-	ngapPdu, err = ngap.Decoder(recvMsg[:n])
-	require.Nil(t, err)
-	assert.True(t, ngapPdu.Present == ngapType.NGAPPDUPresentInitiatingMessage &&
-		ngapPdu.InitiatingMessage.ProcedureCode.Value == ngapType.ProcedureCodeInitialContextSetup,
-		"No InitialContextSetup received.")
-
-	// send ngap Initial Context Setup Response Msg
-	sendMsg, err = GetInitialContextSetupResponse(ue.AmfUeNgapId, ue.RanUeNgapId)
-	require.Nil(t, err)
-	_, err = conn2.Write(sendMsg)
-	require.Nil(t, err)
-
-	// ue send NAS Registration Complete Msg
-	pdu = nasTestpacket.GetRegistrationComplete(nil)
+	pdu = nasTestpacket.GetUlNasTransport_PduSessionEstablishmentRequest(10, nasMessage.ULNASTransportRequestTypeInitialRequest, "internet", &sNssai)
 	pdu, err = EncodeNasPduWithSecurity(ue, pdu, nas.SecurityHeaderTypeIntegrityProtectedAndCiphered, true, false)
 	assert.Nil(t, err)
 	sendMsg, err = GetUplinkNASTransport(ue.AmfUeNgapId, ue.RanUeNgapId, pdu)
 	assert.Nil(t, err)
-	_, err = conn2.Write(sendMsg)
+	_, err = conn.Write(sendMsg)
+	assert.Nil(t, err)
+
+	// receive 12. NGAP-PDU Session Resource Setup Request(DL nas transport((NAS msg-PDU session setup Accept)))
+	n, err = conn.Read(recvMsg)
+	assert.Nil(t, err)
+	ngapPdu, err = ngap.Decoder(recvMsg[:n])
+	assert.Nil(t, err)
+	assert.True(t, ngapPdu.Present == ngapType.NGAPPDUPresentInitiatingMessage &&
+		ngapPdu.InitiatingMessage.ProcedureCode.Value == ngapType.ProcedureCodePDUSessionResourceSetup,
+		"No PDUSessionResourceSetup received.")
+
+	// send 14. NGAP-PDU Session Resource Setup Response
+	sendMsg, err = GetPDUSessionResourceSetupResponse(10, ue.AmfUeNgapId, ue.RanUeNgapId, ranN3Ipv4Addr)
+	assert.Nil(t, err)
+	_, err = conn.Write(sendMsg)
+	assert.Nil(t, err)
+
+	// wait 1s
+	time.Sleep(1 * time.Second)
+
+	sNssai = models.Snssai{
+		Sst: 1,
+		Sd:  "112233",
+	}
+	pdu = nasTestpacket.GetUlNasTransport_PduSessionEstablishmentRequest(11, nasMessage.ULNASTransportRequestTypeInitialRequest, "internet2", &sNssai)
+
+	pdu, err = EncodeNasPduWithSecurity(ue, pdu, nas.SecurityHeaderTypeIntegrityProtectedAndCiphered, true, false)
+	assert.Nil(t, err)
+	sendMsg, err = GetUplinkNASTransport(ue.AmfUeNgapId, ue.RanUeNgapId, pdu)
+	assert.Nil(t, err)
+	_, err = conn.Write(sendMsg)
+	assert.Nil(t, err)
+
+	// receive 12. NGAP-PDU Session Resource Setup Request(DL nas transport((NAS msg-PDU session setup Accept)))
+	n, err = conn.Read(recvMsg)
+	assert.Nil(t, err)
+	ngapPdu, err = ngap.Decoder(recvMsg[:n])
+	assert.Nil(t, err)
+	assert.True(t, ngapPdu.Present == ngapType.NGAPPDUPresentInitiatingMessage &&
+		ngapPdu.InitiatingMessage.ProcedureCode.Value == ngapType.ProcedureCodePDUSessionResourceSetup,
+		"No PDUSessionResourceSetup received.")
+
+	// send 14. NGAP-PDU Session Resource Setup Response
+	sendMsg, err = GetPDUSessionResourceSetupResponse(11, ue.AmfUeNgapId, ue.RanUeNgapId, ranN3Ipv4Addr)
+	assert.Nil(t, err)
+	_, err = conn.Write(sendMsg)
+	assert.Nil(t, err)
+
+	udpServer := ranN3Ipv4Addr + ":2152"
+	udpListener, err := net.ListenPacket("udp", udpServer)
+	if err != nil {
+		fmt.Println(err.Error())
+	}
+	defer udpListener.Close()
+
+	fmt.Printf("UDP server start and listening on %s.\n", udpServer)
+	go func() {
+		for {
+			buf := make([]byte, 1024)
+			_, _, err := udpListener.ReadFrom(buf)
+			if err != nil {
+				continue
+			}
+		}
+	}()
+
+	// wait 1s
+	time.Sleep(1 * time.Second)
+
+	// Send the dummy packet
+	// ping IP(tunnel IP) from 10.60.0.2(127.0.0.1) to 10.60.0.20(127.0.0.8)
+	gtpHdr, err := hex.DecodeString("32ff00340000000100000000")
+	assert.Nil(t, err)
+	icmpData, err := hex.DecodeString("8c870d0000000000101112131415161718191a1b1c1d1e1f202122232425262728292a2b2c2d2e2f3031323334353637")
+	assert.Nil(t, err)
+
+	ipv4hdr := ipv4.Header{
+		Version:  4,
+		Len:      20,
+		Protocol: 1,
+		Flags:    0,
+		TotalLen: 48,
+		TTL:      64,
+		Src:      net.ParseIP("10.60.0.1").To4(),
+		Dst:      net.ParseIP("10.60.0.101").To4(),
+		ID:       1,
+	}
+	checksum := CalculateIpv4HeaderChecksum(&ipv4hdr)
+	ipv4hdr.Checksum = int(checksum)
+
+	v4HdrBuf, err := ipv4hdr.Marshal()
+	assert.Nil(t, err)
+	tt := append(gtpHdr, v4HdrBuf...)
+
+	m := icmp.Message{
+		Type: ipv4.ICMPTypeEcho, Code: 0,
+		Body: &icmp.Echo{
+			ID: 12394, Seq: 1,
+			Data: icmpData,
+		},
+	}
+	b, err := m.Marshal(nil)
+	assert.Nil(t, err)
+	b[2] = 0xaf
+	b[3] = 0x88
+	_, err = upfConn.Write(append(tt, b...))
+	assert.Nil(t, err)
+
+	time.Sleep(1 * time.Second)
+
+	upfConn2, err := ConnectToUpf(ranN3Ipv4Addr, "10.200.200.102", 3002, 2152)
+	assert.Nil(t, err, err)
+	//
+	// Send the dummy packet
+	// ping IP(tunnel IP) from 10.60.0.2(127.0.0.1) to 10.60.0.20(127.0.0.8)
+	gtpHdr, err = hex.DecodeString("32ff00340000000100000000")
+	assert.Nil(t, err)
+	icmpData, err = hex.DecodeString("8c870d0000000000101112131415161718191a1b1c1d1e1f202122232425262728292a2b2c2d2e2f3031323334353637")
+	assert.Nil(t, err)
+
+	ipv4hdr = ipv4.Header{
+		Version:  4,
+		Len:      20,
+		Protocol: 1,
+		Flags:    0,
+		TotalLen: 48,
+		TTL:      64,
+		Src:      net.ParseIP("10.62.0.1").To4(),
+		Dst:      net.ParseIP("10.60.0.102").To4(),
+		ID:       1,
+	}
+	checksum = CalculateIpv4HeaderChecksum(&ipv4hdr)
+	ipv4hdr.Checksum = int(checksum)
+
+	v4HdrBuf, err = ipv4hdr.Marshal()
+	assert.Nil(t, err)
+	tt = append(gtpHdr, v4HdrBuf...)
+
+	m = icmp.Message{
+		Type: ipv4.ICMPTypeEcho, Code: 0,
+		Body: &icmp.Echo{
+			ID: 12394, Seq: 1,
+			Data: icmpData,
+		},
+	}
+	b, err = m.Marshal(nil)
+
+	assert.Nil(t, err)
+	b[2] = 0xaf
+	b[3] = 0x88
+	_, err = upfConn2.Write(append(tt, b...))
 	assert.Nil(t, err)
 
 	time.Sleep(1 * time.Second)
@@ -683,8 +740,9 @@ func TestMultiAmfRegistration(t *testing.T) {
 	DelAccessAndMobilitySubscriptionDataFromMongoDB(ue.Supi, servingPlmnId)
 	DelSmfSelectionSubscriptionDataFromMongoDB(ue.Supi, servingPlmnId)
 
-	// close Connections
-	conn2.Close()
+	// close Connection
+	conn.Close()
 
+	// terminate all NF
 	NfTerminate()
 }
