+++ test/reSynchronization_test.go
@@ -1,0 +1,313 @@
+package test
+
+import (
+	"bytes"
+	"encoding/binary"
+	"encoding/hex"
+	"fmt"
+	"net"
+	"testing"
+	"time"
+
+	"github.com/free5gc/nas"
+	"github.com/free5gc/nas/nasMessage"
+	"github.com/free5gc/nas/nasType"
+	"github.com/free5gc/nas/security"
+	"github.com/free5gc/ngap"
+	"github.com/free5gc/openapi/models"
+	"github.com/free5gc/test/consumerTestdata/UDM/TestGenAuthData"
+	"github.com/free5gc/test/nasTestpacket"
+	"github.com/free5gc/util/milenage"
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
+	"golang.org/x/net/icmp"
+	"golang.org/x/net/ipv4"
+)
+
+func TestReSynchronization(t *testing.T) {
+	var n int
+	var sendMsg []byte
+	var recvMsg = make([]byte, 2048)
+
+	// RAN connect to AMF
+	conn, err := ConnectToAmf(amfN2Ipv4Addr, ranN2Ipv4Addr, 38412, 9487)
+	assert.Nil(t, err)
+
+	// RAN connect to UPF
+	upfConn, err := ConnectToUpf(ranN3Ipv4Addr, "10.200.200.102", 2152, 2152)
+	assert.Nil(t, err)
+
+	// send NGSetupRequest Msg
+	sendMsg, err = GetNGSetupRequest([]byte("\x00\x01\x02"), 24, "free5gc")
+	assert.Nil(t, err)
+	_, err = conn.Write(sendMsg)
+	assert.Nil(t, err)
+
+	// receive NGSetupResponse Msg
+	n, err = conn.Read(recvMsg)
+	assert.Nil(t, err)
+	_, err = ngap.Decoder(recvMsg[:n])
+	assert.Nil(t, err)
+
+	// New UE
+	// ue := NewRanUeContext("imsi-2089300007487", 1, security.AlgCiphering128NEA2, security.AlgIntegrity128NIA2, models.AccessType__3_GPP_ACCESS)
+	ue := NewRanUeContext("imsi-2089300007487", 1, security.AlgCiphering128NEA0, security.AlgIntegrity128NIA2,
+		models.AccessType__3_GPP_ACCESS)
+	ue.AmfUeNgapId = 1
+	ue.AuthenticationSubs = GetAuthSubscription(TestGenAuthData.MilenageTestSet19.K,
+		TestGenAuthData.MilenageTestSet19.OPC,
+		TestGenAuthData.MilenageTestSet19.OP)
+	// insert UE data to MongoDB
+
+	servingPlmnId := "20893"
+	insertUeToMongoDB(t, ue, servingPlmnId)
+
+	// send InitialUeMessage(Registration Request)(imsi-2089300007487)
+	mobileIdentity5GS := nasType.MobileIdentity5GS{
+		Len:    12, // suci
+		Buffer: []uint8{0x01, 0x02, 0xf8, 0x39, 0xf0, 0xff, 0x00, 0x00, 0x00, 0x00, 0x47, 0x78},
+	}
+
+	ueSecurityCapability := ue.GetUESecurityCapability()
+	registrationRequest := nasTestpacket.GetRegistrationRequest(
+		nasMessage.RegistrationType5GSInitialRegistration, mobileIdentity5GS, nil, ueSecurityCapability, nil, nil, nil)
+	sendMsg, err = GetInitialUEMessage(ue.RanUeNgapId, registrationRequest, "")
+	assert.Nil(t, err)
+	_, err = conn.Write(sendMsg)
+	assert.Nil(t, err)
+
+	// receive NAS Authentication Request Msg
+	n, err = conn.Read(recvMsg)
+	assert.Nil(t, err)
+	ngapMsg, err := ngap.Decoder(recvMsg[:n])
+	assert.Nil(t, err)
+
+	nasPdu := GetNasPdu(ue, ngapMsg.InitiatingMessage.Value.DownlinkNASTransport)
+	require.NotNil(t, nasPdu)
+	require.NotNil(t, nasPdu.GmmMessage, "GMM message is nil")
+	require.Equal(t, nasPdu.GmmHeader.GetMessageType(), nas.MsgTypeAuthenticationRequest,
+		"Received wrong GMM message. Expected Authentication Request.")
+
+	// gen AK
+	K, OPC := make([]byte, 16), make([]byte, 16)
+	K, _ = hex.DecodeString(ue.AuthenticationSubs.PermanentKey.PermanentKeyValue)
+	OPC, _ = hex.DecodeString(ue.AuthenticationSubs.Opc.OpcValue)
+	SQN := make([]byte, 6)
+	AK := make([]byte, 6)
+
+	rand := nasPdu.AuthenticationRequest.GetRANDValue()
+	milenage.F2345(OPC, K, rand[:], nil, nil, nil, AK, nil)
+	autn := nasPdu.AuthenticationRequest.GetAUTN()
+	SQNxorAK := autn[:6]
+	for i := 0; i < 6; i++ {
+		SQN[i] = AK[i] ^ SQNxorAK[i]
+	}
+	const SqnMAx int64 = 0x7FFFFFFFFFF
+	const SqnMs int64 = 0
+	const IND int64 = 32
+	var newSqnMsString string
+	SQNBuffer := make([]byte, 8)
+	copy(SQNBuffer[2:], SQN)
+	r := bytes.NewReader(SQNBuffer)
+	var retrieveSqn int64
+	if err := binary.Read(r, binary.BigEndian, &retrieveSqn); err != nil {
+		fmt.Println("err", err)
+		return
+	}
+
+	delita := retrieveSqn - SqnMs
+	if delita < 0x7FFFFFFFFFF {
+		newSqnMsString = "000000000000"
+	}
+
+	newSqnMs, _ := hex.DecodeString(newSqnMsString)
+	MAC_A, MAC_S := make([]byte, 8), make([]byte, 8)
+	CK, IK := make([]byte, 16), make([]byte, 16)
+	RES := make([]byte, 8)
+	AK, AKstar := make([]byte, 6), make([]byte, 6)
+	AMF, _ := hex.DecodeString("0000")
+	milenage.F1(OPC, K, rand[:], newSqnMs, AMF, MAC_A, MAC_S)
+	milenage.F2345(OPC, K, rand[:], RES, CK, IK, AK, AKstar)
+
+	SQNmsxorAK := make([]byte, 6)
+	for i := 0; i < len(SQN); i++ {
+		SQNxorAK[i] = SQN[i] ^ AK[i]
+	}
+	ColSQNmsxorAK := make([]byte, 6)
+	for i := 0; i < len(SQN); i++ {
+		ColSQNmsxorAK[i] = SQNmsxorAK[i] ^ AKstar[i]
+	}
+	AUTS := append(ColSQNmsxorAK, MAC_S...)
+	// compute SQN by AUTN, K, AK
+	// suppose
+	// send NAS Authentication Rejcet
+	// failureParam := []uint8{0x68, 0x58, 0x15, 0x86, 0x1f, 0xec, 0x0f, 0xa9, 0x48, 0xe8, 0xb2, 0x3a, 0x08, 0x62}
+	failureParam := AUTS
+	pdu := nasTestpacket.GetAuthenticationFailure(nasMessage.Cause5GMMSynchFailure, failureParam)
+	sendMsg, err = GetUplinkNASTransport(ue.AmfUeNgapId, ue.RanUeNgapId, pdu)
+	assert.Nil(t, err)
+	_, err = conn.Write(sendMsg)
+	assert.Nil(t, err)
+
+	// receive NAS Authentication Request Msg
+	n, err = conn.Read(recvMsg)
+	assert.Nil(t, err)
+	ngapMsg, err = ngap.Decoder(recvMsg[:n])
+	assert.Nil(t, err)
+
+	// Calculate for RES*
+	nasPdu = GetNasPdu(ue, ngapMsg.InitiatingMessage.Value.DownlinkNASTransport)
+	require.NotNil(t, nasPdu)
+	require.NotNil(t, nasPdu.GmmMessage, "GMM message is nil")
+	require.Equal(t, nasPdu.GmmHeader.GetMessageType(), nas.MsgTypeAuthenticationRequest,
+		"Received wrong GMM message. Expected Authentication Request.")
+	rand = nasPdu.AuthenticationRequest.GetRANDValue()
+
+	milenage.F2345(OPC, K, rand[:], nil, nil, nil, AK, nil)
+	autn = nasPdu.AuthenticationRequest.GetAUTN()
+	SQNxorAK = autn[:6]
+
+	for i := 0; i < 6; i++ {
+		SQN[i] = AK[i] ^ SQNxorAK[i]
+	}
+	fmt.Printf("retrieve SQN %x\n", SQN)
+	ue.AuthenticationSubs.SequenceNumber = hex.EncodeToString(SQN)
+	resStar := ue.DeriveRESstarAndSetKey(ue.AuthenticationSubs, rand[:], "5G:mnc093.mcc208.3gppnetwork.org")
+
+	// send NAS Authentication Response
+	pdu = nasTestpacket.GetAuthenticationResponse(resStar, "")
+	sendMsg, err = GetUplinkNASTransport(ue.AmfUeNgapId, ue.RanUeNgapId, pdu)
+	assert.Nil(t, err)
+	_, err = conn.Write(sendMsg)
+	assert.Nil(t, err)
+
+	// receive NAS Security Mode Command Msg
+	n, err = conn.Read(recvMsg)
+	assert.Nil(t, err)
+	ngapPdu, err := ngap.Decoder(recvMsg[:n])
+	require.Nil(t, err)
+	require.NotNil(t, ngapPdu)
+	nasPdu = GetNasPdu(ue, ngapPdu.InitiatingMessage.Value.DownlinkNASTransport)
+	require.NotNil(t, nasPdu)
+	require.NotNil(t, nasPdu.GmmMessage, "GMM message is nil")
+	require.Equal(t, nasPdu.GmmHeader.GetMessageType(), nas.MsgTypeSecurityModeCommand,
+		"Received wrong GMM message. Expected Security Mode Command.")
+
+	// send NAS Security Mode Complete Msg
+	registrationRequestWith5GMM := nasTestpacket.GetRegistrationRequest(nasMessage.RegistrationType5GSInitialRegistration,
+		mobileIdentity5GS, nil, ueSecurityCapability, ue.Get5GMMCapability(), nil, nil)
+	pdu = nasTestpacket.GetSecurityModeComplete(registrationRequestWith5GMM)
+	pdu, err = EncodeNasPduWithSecurity(ue, pdu, nas.SecurityHeaderTypeIntegrityProtectedAndCipheredWithNew5gNasSecurityContext, true, true)
+	assert.Nil(t, err)
+	sendMsg, err = GetUplinkNASTransport(ue.AmfUeNgapId, ue.RanUeNgapId, pdu)
+	assert.Nil(t, err)
+	_, err = conn.Write(sendMsg)
+	assert.Nil(t, err)
+
+	// receive ngap Initial Context Setup Request Msg
+	n, err = conn.Read(recvMsg)
+	assert.Nil(t, err)
+	_, err = ngap.Decoder(recvMsg[:n])
+	assert.Nil(t, err)
+
+	// send ngap Initial Context Setup Response Msg
+	sendMsg, err = GetInitialContextSetupResponse(ue.AmfUeNgapId, ue.RanUeNgapId)
+	assert.Nil(t, err)
+	_, err = conn.Write(sendMsg)
+	assert.Nil(t, err)
+
+	// send NAS Registration Complete Msg
+	pdu = nasTestpacket.GetRegistrationComplete(nil)
+	pdu, err = EncodeNasPduWithSecurity(ue, pdu, nas.SecurityHeaderTypeIntegrityProtectedAndCiphered, true, false)
+	assert.Nil(t, err)
+	sendMsg, err = GetUplinkNASTransport(ue.AmfUeNgapId, ue.RanUeNgapId, pdu)
+	assert.Nil(t, err)
+	_, err = conn.Write(sendMsg)
+	assert.Nil(t, err)
+
+	// receive UE Configuration Update Command Msg
+	recvUeConfigUpdateCmd(t, recvMsg, conn)
+
+	time.Sleep(100 * time.Millisecond)
+
+	// send GetPduSessionEstablishmentRequest Msg
+	sNssai := models.Snssai{
+		Sst: 1,
+		Sd:  "fedcba",
+	}
+	pdu = nasTestpacket.GetUlNasTransport_PduSessionEstablishmentRequest(10, nasMessage.ULNASTransportRequestTypeInitialRequest, "internet", &sNssai)
+	pdu, err = EncodeNasPduWithSecurity(ue, pdu, nas.SecurityHeaderTypeIntegrityProtectedAndCiphered, true, false)
+	assert.Nil(t, err)
+	sendMsg, err = GetUplinkNASTransport(ue.AmfUeNgapId, ue.RanUeNgapId, pdu)
+	assert.Nil(t, err)
+	_, err = conn.Write(sendMsg)
+	assert.Nil(t, err)
+
+	// receive 12. NGAP-PDU Session Resource Setup Request(DL nas transport((NAS msg-PDU session setup Accept)))
+	n, err = conn.Read(recvMsg)
+	assert.Nil(t, err)
+	_, err = ngap.Decoder(recvMsg[:n])
+	assert.Nil(t, err)
+
+	// send 14. NGAP-PDU Session Resource Setup Response
+	sendMsg, err = GetPDUSessionResourceSetupResponse(10, ue.AmfUeNgapId, ue.RanUeNgapId, ranN3Ipv4Addr)
+	assert.Nil(t, err)
+	_, err = conn.Write(sendMsg)
+	assert.Nil(t, err)
+
+	// wait 1s
+	time.Sleep(1 * time.Second)
+
+	// Send the dummy packet
+	// ping IP(tunnel IP) from 10.60.0.2(127.0.0.1) to 10.60.0.20(127.0.0.8)
+	gtpHdr, err := hex.DecodeString("32ff00340000000100000000")
+	assert.Nil(t, err)
+	icmpData, err := hex.DecodeString("8c870d0000000000101112131415161718191a1b1c1d1e1f202122232425262728292a2b2c2d2e2f3031323334353637")
+	assert.Nil(t, err)
+
+	ipv4hdr := ipv4.Header{
+		Version:  4,
+		Len:      20,
+		Protocol: 1,
+		Flags:    0,
+		TotalLen: 48,
+		TTL:      64,
+		Src:      net.ParseIP("10.60.0.1").To4(),
+		Dst:      net.ParseIP("10.60.0.101").To4(),
+		ID:       1,
+	}
+	checksum := CalculateIpv4HeaderChecksum(&ipv4hdr)
+	ipv4hdr.Checksum = int(checksum)
+
+	v4HdrBuf, err := ipv4hdr.Marshal()
+	assert.Nil(t, err)
+	tt := append(gtpHdr, v4HdrBuf...)
+	assert.Nil(t, err)
+
+	m := icmp.Message{
+		Type: ipv4.ICMPTypeEcho, Code: 0,
+		Body: &icmp.Echo{
+			ID: 12394, Seq: 1,
+			Data: icmpData,
+		},
+	}
+	b, err := m.Marshal(nil)
+	assert.Nil(t, err)
+	b[2] = 0xaf
+	b[3] = 0x88
+	_, err = upfConn.Write(append(tt, b...))
+	assert.Nil(t, err)
+
+	time.Sleep(1 * time.Second)
+
+	// delete test data
+	DelAuthSubscriptionToMongoDB(ue.Supi)
+	DelAccessAndMobilitySubscriptionDataFromMongoDB(ue.Supi, servingPlmnId)
+	DelSmfSelectionSubscriptionDataFromMongoDB(ue.Supi, servingPlmnId)
+
+	// close Connection
+	conn.Close()
+
+	// terminate all NF
+	NfTerminate()
+}
